requirements.txt
scikit-learn
fastapi
uvicorn
pydantic
shap
pandas
numpy
pytest
pytest-cov
httpx
sklearn-genetic-opt
deap
matplotlib
